Adapt Bootstrap 5 SCSS for Bootstrap 5.3+
The existing code already uses calc() in some places, so using it instead of add() is safe.

Because Bootstrap 5.3 is using more and more CSS variables, we need to use color-mix() which has an good evergreen browser support : https://caniuse.com/?search=color-mix

To maintain Bootstrap 5.0 to 5.2 compatibility for older browsers, a separate "tom-select.bootstrap5.3.scss" file should be create with this code.

There was an SASS deprecation warning with Bootstrap 4 that's also fixed.